requirements.txt
requests
openai

# openai library
openai

# http client
requests

# simply ui framework
streamlit

# .env utility
python-dotenv

# jupyter notbook
jupyter

#pdf library
pypdf2
pdfplumber
camelot-py[base] 
pandas 
tabulate
pyhwp
